*	UPDATED:    Translations and added some partials (Catalan, Persian) *	FIX:        for cleanup cron running on non-existent tables.
git-svn-id: https://plugins.svn.wordpress.org/wp-simple-firewall/trunk@884735 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ src/icwp-processor-commentsfilter.php
@@ -438,6 +438,9 @@ class ICWP_CommentsFilterProcessor_V1 extends ICWP_BaseDbProcessor_WPSF {
 	 * It'll delete everything older than 24hrs.
 	 */
 	public function cleanupDatabase() {
+		if ( !$this->getTableExists() ) {
+			return;
+		}
 		$nTimeStamp = time() - DAY_IN_SECONDS;
 		$this->deleteAllRowsOlderThan( $nTimeStamp );
 	}
